Rename BankingApi to TppBankingApi
 #83
