--- opba-banking-rest-api/src/main/java/de/adorsys/opba/bankingapi/search/resource/TppBankSearchResource.java
@@ -1,7 +1,0 @@
-package de.adorsys.opba.bankingapi.search.resource;
-
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class TppBankSearchResource implements TppBankSearchApi {
-}
