--- opba-banking-rest-api/src/main/java/de/adorsys/opba/bankingapi/token/resource/TppTokenResource.java
@@ -1,7 +1,0 @@
-package de.adorsys.opba.bankingapi.token.resource;
-
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class TppTokenResource implements TppTokenApi {
-}
