+++ opba-banking-rest-api/src/main/java/de/adorsys/opba/tppbankingapi/config/BankingApiConfig.java
@@ -1,9 +1,9 @@
-package de.adorsys.opba.bankingapi.config;
+package de.adorsys.opba.tppbankingapi.config;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-@ComponentScan("de.adorsys.opba.bankingapi")
+@ComponentScan("de.adorsys.opba.tppbankingapi")
 public class BankingApiConfig {
 }
